Shadow Bug mejorado
Problema que no siempre se apagaba la sombra, arreglado
