+++ Assets/[Scripts]/Flashlight.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0e4f3070d1df7974c8621cbe5842a8c8
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
