+++ Assets/[Scripts]/Flashlight/FlashlightDetector.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 663a7982488f6554ab15912f9fadda3a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
